Moved dashboard video logic to history
